refactor: extract benchmark boundary cli commands
