+++ loopx/cli_commands/__init__.py
@@ -8,6 +8,10 @@ Command modules expose two small functions:
 The top-level CLI keeps global options, registry fallback, and dispatch order.
 """
 
+from .benchmark_boundary import (
+    handle_benchmark_boundary_command,
+    register_benchmark_boundary_commands,
+)
 from .doctor import handle_doctor_command, register_doctor_command
 from .dreaming import handle_dreaming_command, register_dreaming_commands
 from .history import handle_history_command, register_history_command
@@ -44,6 +48,7 @@ from .status import (
 from .todo import handle_todo_command, register_todo_command
 
 __all__ = [
+    "handle_benchmark_boundary_command",
     "handle_check_command",
     "handle_codex_cli_bounded_visible_pilot_adapter_command",
     "handle_codex_cli_bootstrap_message_command",
@@ -72,6 +77,7 @@ __all__ = [
     "handle_review_packet_command",
     "handle_status_command",
     "handle_todo_command",
+    "register_benchmark_boundary_commands",
     "register_doctor_command",
     "register_dreaming_commands",
     "register_history_command",
